+++ src/commands/index.ts
@@ -9,3 +9,4 @@ export * from './s3';
 export * from './gcp';
 export * from './azure';
 export * from './security';
+export * from './list';